requirements.txt
# Web3 and Blockchain
web3==5.24.0
eth-account==0.5.6
eth-utils==2.1.0
eth-typing==4.0.0

# FastAPI and Web
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-jose==3.3.0
passlib==1.7.4
pydantic==1.8.2
pydantic-settings==2.0.3
bcrypt==3.2.0

# Database
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
alembic==1.7.1
asyncpg==0.29.0

# AI and ML
torch==2.1.1
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
tensorflow==2.14.0
joblib==1.0.1
matplotlib==3.4.3
seaborn==0.11.2
python-dateutil==2.8.2
pytz==2021.1

# Robotics
ros2==0.0.1
rclpy==3.8.0
geometry-msgs==4.13.0
sensor-msgs==4.3.0

# Monitoring and Logging
sentry-sdk==1.32.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Testing
pytest==6.2.5
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.2

# Utilities
python-dotenv==0.19.0
requests==2.26.0
aiohttp==3.9.1
tenacity==8.2.3
cachetools==5.3.2

# AWS and Azure
boto3==1.29.3
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.7
mkdocstrings==0.22.0 